# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Event-Driven Notification API
  version: 1.0.0
paths:
  /notifications:
    post:
      summary: Create a batch of notifications
      parameters:
        - in: header
          name: X-Correlation-ID
          schema:
            type: string
        - in: header
          name: Idempotency-Key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBatchRequest'
      responses:
        '200':
          description: Existing idempotent batch returned
        '201':
          description: Created
        '400':
          description: Invalid request
    get:
      summary: List notifications
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, queued, sent, failed, cancelled]
        - in: query
          name: channel
          schema:
            type: string
            enum: [sms, email, push]
        - in: query
          name: batch_id
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Notification list
  /notifications/{id}:
    get:
      summary: Get notification status by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification
        '404':
          description: Not found
    delete:
      summary: Cancel a pending notification
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cancelled
        '409':
          description: Notification not found or not cancellable
  /templates:
    post:
      summary: Create a notification template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateCreate'
      responses:
        '201':
          description: Created
    get:
      summary: List templates
      responses:
        '200':
          description: Template list
  /templates/{id}:
    get:
      summary: Get template by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template
        '404':
          description: Not found
  /batches/{batch_id}/notifications:
    get:
      summary: Query notifications by batch ID
      parameters:
        - in: path
          name: batch_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch notifications
  /metrics:
    get:
      summary: Real-time system metrics
      responses:
        '200':
          description: Metrics snapshot
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Healthy
components:
  schemas:
    CreateBatchRequest:
      type: object
      required: [notifications]
      properties:
        notifications:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/NotificationCreate'
    NotificationCreate:
      type: object
      required: [recipient, channel]
      properties:
        recipient:
          type: string
        channel:
          type: string
          enum: [sms, email, push]
        content:
          type: string
        template_id:
          type: string
        template_data:
          type: object
          additionalProperties:
            type: string
        scheduled_at:
          type: string
          format: date-time
        priority:
          type: string
          enum: [high, normal, low]
          default: normal
    TemplateCreate:
      type: object
      required: [name, channel, content]
      properties:
        name:
          type: string
        channel:
          type: string
          enum: [sms, email, push]
        content:
          type: string